Attributes working
Permission flags can be set to prevent users from using service methods
on documents.
- create, patch and update calls without adequate permissions will
serve errors back to the user
- find calls will filter unaccessible documents out of results
- get calls will throw errors if documents are inaccessible

Further flags can be set on individual data parameters.
- specific fields can be prevented from being edited
- specific fields can be filtered out of find/get calls

Permissions can also be configured with functions instead of string
flags, but this needs to be tested and the API properly figured out.